+++ board/safety_declarations.h
@@ -84,7 +84,7 @@ void generic_rx_checks(bool stock_ecu_detected);
 void relay_malfunction_set(void);
 void relay_malfunction_reset(void);
 
-typedef const addr_checks* (*safety_hook_init)(uint32_t param);
+typedef const addr_checks* (*safety_hook_init)(uint16_t param);
 typedef int (*rx_hook)(CANPacket_t *to_push);
 typedef int (*tx_hook)(CANPacket_t *to_send, bool longitudinal_allowed);
 typedef int (*tx_lin_hook)(int lin_num, uint8_t *data, int len);
